citi-backend/requirements.txt
django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
channels==4.1.0
channels-redis==4.2.0
celery==5.4.0
redis==5.0.8
anthropic==0.34.2
dj-database-url==2.2.0
python-decouple==3.8
psycopg2-binary==2.9.9
whitenoise==6.7.0
